add paginated and update endpoints to user change endpoint URI add conditions

+++ src/controllers/UserController.js
@@ -9,14 +9,24 @@ const findAllUsers = asyncErrorHandler(async (req, resp, next) => {
     await sendResponse(resp, 200, {users: users, count: users.length});
 });
 
+const findAllUsersWithPagination = asyncErrorHandler(async (req, resp, next) => {
+    const usersWithPagination = await userService.findAllUsersWithPagination(req);
+    await sendResponse(resp, 200, usersWithPagination);
+});
+
+const findAllUsersBySearchWithPagination = asyncErrorHandler(async (req, resp, next) => {
+    const usersWithPagination = await userService.findAllUsersBySearchWithPagination(req);
+    await sendResponse(resp, 200, usersWithPagination);
+});
+
 const findUserById = asyncErrorHandler(async (req, resp, next) => {
     const user = await userService.findUserById(req.params);
     await sendResponse(resp, 200, {user});
 });
 
 const updateUser = asyncErrorHandler(async (req, resp, next) => {
-    if (req.body.password || req.body.confirmPassword) {
-        const error = new BadRequestError('You cannot update password');
+    if (req.params.id === req.user.id) {
+        const error = new BadRequestError('For current user can not change role.');
         return next(error);
     }
 
@@ -24,21 +34,43 @@ const updateUser = asyncErrorHandler(async (req, resp, next) => {
     await sendResponse(resp, 201, {user});
 });
 
-const changePassword = asyncErrorHandler(async (req, resp, next) => {
+const updateUserByAuthUser = asyncErrorHandler(async (req, resp, next) => {
+    if (req.body.password || req.body.confirmPassword) {
+        const error = new BadRequestError('You cannot update password');
+        return next(error);
+    }
+
+    const user = await userService.updateUserByAuthUser(req);
+    await sendResponse(resp, 201, {user});
+});
+
+const changePasswordByAuthUser = asyncErrorHandler(async (req, resp, next) => {
     if (!req.body.password || !req.body.confirmPassword) {
         const error = new BadRequestError('Please provide password & confirmPassword!');
         return next(error);
     }
 
-    const userWithToken = await userService.changeUserPassword(req, resp);
+    const userWithToken = await userService.changeUserPasswordByAuthUser(req, resp);
     await sendResponse(resp, 201, userWithToken);
 });
 
 const deleteUser = asyncErrorHandler(async (req, resp, next) => {
+    if (req.params.id === req.user.id) {
+        const error = new BadRequestError('Current user can not delete the user account.');
+        return next(error);
+    }
+
     const user = await userService.deleteUser(req.params);
     await sendResponse(resp, 204, {id: user._id});
 });
 
 module.exports = {
-    findAllUsers, findUserById, updateUser, changePassword, deleteUser
+    findAllUsers,
+    findAllUsersWithPagination,
+    findAllUsersBySearchWithPagination,
+    findUserById,
+    updateUser,
+    updateUserByAuthUser,
+    changePasswordByAuthUser,
+    deleteUser
 }
